Rotate 0 degrees to midnight via svg <g transform>
Instead of coding each individual angle setting to set 0 to midnight,
just wrap the whole clock face in a <g> element to rotate everything.

This simplifies the construction of the individual elements,
and fixes a bug where some things were getting rotated 90 degrees
counter-clockwise twice (the keys and the tickmarks).