Add system service template.
This is just a checkpoint of the furthest I took the udev rules +
SYSTEMD_WANTS path of exploration.  The mechanism is complicated and
hard to configure and install.  I also did not have any luck getting
BindTo to allow replugs to stop and restart the service appropriately.

Instead, I'll next explore using udevlib through pyudev.